Move ChatList, move and rename HomeViewModel
