+++ app/src/androidTest/java/com/google/android/samples/socialite/ui/home/ChatListViewModelTest.kt
@@ -20,17 +20,18 @@ import androidx.test.ext.junit.runners.AndroidJUnit4
 import app.cash.turbine.test
 import com.google.android.samples.socialite.awaitNotEmpty
 import com.google.android.samples.socialite.repository.createTestRepository
+import com.google.android.samples.socialite.ui.home.chatlist.ChatListViewModel
 import com.google.common.truth.Truth.assertThat
 import kotlinx.coroutines.test.runTest
 import org.junit.Test
 import org.junit.runner.RunWith
 
 @RunWith(AndroidJUnit4::class)
-class HomeViewModelTest {
+class ChatListViewModelTest {
 
     @Test
     fun initialize() = runTest {
-        val viewModel = HomeViewModel(createTestRepository())
+        val viewModel = ChatListViewModel(createTestRepository())
         viewModel.chats.test {
             assertThat(awaitNotEmpty()).hasSize(4)
         }
