+++ app/src/main/java/com/google/android/samples/socialite/widget/SociaLiteAppWidgetConfigActivity.kt
@@ -45,14 +45,12 @@ import androidx.compose.ui.draw.clip
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import androidx.hilt.navigation.compose.hiltViewModel
 import com.google.android.samples.socialite.di.AppCoroutineScope
 import com.google.android.samples.socialite.model.Contact
 import com.google.android.samples.socialite.repository.ChatRepository
 import com.google.android.samples.socialite.ui.SocialTheme
 import com.google.android.samples.socialite.ui.home.HomeAppBar
 import com.google.android.samples.socialite.ui.home.HomeBackground
-import com.google.android.samples.socialite.ui.home.HomeViewModel
 import com.google.android.samples.socialite.ui.rememberIconPainter
 import com.google.android.samples.socialite.widget.model.WidgetModelRepository
 import dagger.hilt.android.AndroidEntryPoint
@@ -92,7 +90,6 @@ class SociaLiteAppWidgetConfigActivity : ComponentActivity() {
                 ) { innerPadding ->
 
                     HomeBackground()
-                    val viewModel: HomeViewModel = hiltViewModel()
                     LazyColumn(
                         modifier = modifier,
                         contentPadding = innerPadding,
